+++ backend/config/app.js
@@ -1,0 +1,10 @@
+const mongoose = require('mongoose'),
+    express = require('./express');
+
+module.exports.start = function() {
+    const app = express.init();
+    const port = process.env.PORT || 8080;
+    app.listen(port, () => 
+        console.log(`Server now running on port `, port)
+    );
+};
