+++ backend/config/passport.js
@@ -2,7 +2,6 @@ const JwtStrategy = require("passport-jwt").Strategy,
     ExtractJwt = require("passport-jwt").ExtractJwt,
     mongoose = require("mongoose"),
     Student = require("../models/students.model");
-    // keys = require("./config");
 
 const opts = {};
 opts.jwtFromRequest = ExtractJwt.fromAuthHeaderAsBearerToken();
