+++ backend/controllers/email.controller.js
@@ -1,8 +1,7 @@
 const nodemailer = require('nodemailer');
-// const creds = require('../config/config.js');
 
 function body(userId){
-  const link = /*'localhost:3000/homepage/'*/ 'https://simmons-data-vis.herokuapp.com/homepage/'+ userId ;
+  const link = 'https://simmons-data-vis.herokuapp.com/homepage/' + userId ;
 
   return "Hi, to conduct your survey please follow this link: " + link;
 }
