+++ backend/server.js
@@ -1,0 +1,2 @@
+const app = require('./config/app'),
+    server = app.start();
